requirements.txt
transformers
numpy
torch
hydra-core
tqdm
fschat[model_worker,webui]
